Baby steps, but *correct* baby steps
